feat: Attempt to rename the service to articles
